+++ android/app/src/main/java/com/yagubogu/data/datasource/stats/StatsRemoteDataSource.kt
@@ -1,9 +1,9 @@
-package com.yagubogu.data.datasource
+package com.yagubogu.data.datasource.stats
 
 import com.yagubogu.data.dto.response.AverageStatisticResponse
-import com.yagubogu.data.dto.response.StatsCountsResponse
-import com.yagubogu.data.dto.response.StatsLuckyStadiumsResponse
-import com.yagubogu.data.dto.response.StatsWinRateResponse
+import com.yagubogu.data.dto.response.stats.StatsCountsResponse
+import com.yagubogu.data.dto.response.stats.StatsLuckyStadiumsResponse
+import com.yagubogu.data.dto.response.stats.StatsWinRateResponse
 import com.yagubogu.data.service.StatsApiService
 import com.yagubogu.data.util.safeApiCall
