+++ android/app/src/main/java/com/yagubogu/data/dto/response/stats/StatsCountsResponse.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.data.dto.response
+package com.yagubogu.data.dto.response.stats
 
 import com.yagubogu.domain.model.StatsCounts
 import kotlinx.serialization.SerialName
