+++ android/app/src/main/java/com/yagubogu/data/dto/response/stats/StatsLuckyStadiumsResponse.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.data.dto.response
+package com.yagubogu.data.dto.response.stats
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
