+++ android/app/src/main/java/com/yagubogu/data/service/StatsApiService.kt
@@ -1,9 +1,9 @@
 package com.yagubogu.data.service
 
 import com.yagubogu.data.dto.response.AverageStatisticResponse
-import com.yagubogu.data.dto.response.StatsCountsResponse
-import com.yagubogu.data.dto.response.StatsLuckyStadiumsResponse
-import com.yagubogu.data.dto.response.StatsWinRateResponse
+import com.yagubogu.data.dto.response.stats.StatsCountsResponse
+import com.yagubogu.data.dto.response.stats.StatsLuckyStadiumsResponse
+import com.yagubogu.data.dto.response.stats.StatsWinRateResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
